Implement tasks struct with serializing and deserializing functionality
The core of our application is the management of tasks. As part of this
management, the program needs to persist those tasks and later be able
to load them.
With previous changes we have imported Rust's serde library and the
backend for serializing to and deserializing back from JSON. This change
introduces a new modules, tasks.rs, that provides basic task management
including the loading from files.