Introduce model-view-controller architecture
This change introduces the basis for the overall program architecture.
We will decouple the presentation layer, a terminal based user interface
powered by the 'termion' library, from the actual model, i.e., the
already introduced task management code. This functionality is
encapsulated in a view.
An intermediate layer, the controller, will take care of providing the
core business logic that is planned to follow.

As view and controller are abstract entities we have two concrete
implementations in the form of the TermUi and Orchestrator structs. Note
that we restrict our architecture to a single view as of now. Supporting
multiple views increases the complexity of the code and we have no
additional views planned.

Note that we maintain a very loose definition of the MVC architecture.
In particular in many definitions the controller is responsible for
handling input. In our scenario it is the view that handles this job.
This decision was driven by necessity: The termion library provides for
input and output and it powers our view. It makes sense to combine those
responsibilities into one entity. Furthermore, it is unclear how
characters the user just entered, which are meant to be displayed as
they are entered, would be conveyed from the controller to the view in a
concise and intuitive way.