Add support for displaying tasks
Now that we have the basic architecture in place, this change is the
first to make the terminal UI a bit more useful. In particular, it adds
support for displaying the tasks loaded in the model. Not only that, it
also provides the capability for the user to select particular entries,
which then appear in a different color. By doing so we implicitly
support scrolling. That is, if the list of tasks exceeds the size of the
screen the cursor allows for viewing items below or above the current
window being displayed.
The entire UI will be manageable using VIM-style key bindings and so
selecting the task above the current selection can be achieved using 'k'
while selecting the one below is bound to 'j'.