Add support for saving the current set of tasks
This change adds support for saving the current set of tasks back into
the configuration file. As there is no way for the user to add or remove
tasks currently only later patches will make this change truly useful.
In order to indicate the status of the save operation (i.e., success or
failure) a dedicated input/output area is introduced, which will serve
multiple purposes in the future.