Add support for adding new tasks
So far we have not enabled any useful work flows because we merely
allowed for loading an existing set of tasks. That changes with this
patch as it introduces support for adding new tasks. Using the 'a' key
the user is allowed to enter text (via the previously introduced
input/output area) that is treated as the task's summary.