Use pretty printing when storing tasks into files
One of the essential properties of the program is that the configuration
(mostly the task data as of now) is human readable, editable, and
version controllable. JSON provides those benefits by default, however,
in the current form everything is stored on a single line, making it
very hard to handle by traditionally line based program.
This change adjusts the task saving code to pretty print the JSON data
before saving it into a file.