Refactor event handling code
With an ever increasing number of cases required to be covered by the
UI, the complexity of this part of the code is anticipated to grow to
unmanageable levels. For example, we currently implement dispatching
of key press events to the proper receiver by adjusting a function
pointer when our state machine reached a certain state. Although that
works, what we really need is a proper UI that supports focusing of
different widgets, which implicitly adjusts the receiving widget as
well. That simplifies reasoning and keeps the actual widget specific
logic separate from the generic UI functionality.
While such a UI is in the makes, this change prepares for its use in the
program by refactoring the event handling logic. In particular, instead
of having the checking for new events inside the view itself, this
change moves it out to the caller.