Remove return value from View::render
Currently we pass through errors encountered during the rendering path.
Those errors, however, are very specific to the termion backend. In
fact, it is uncommon to make render functions failable in the first
place.
In anticipation that the soon-to-be-used UI library will not support
returning errors from the render path, this change removes the return
value of the view's render method.