Remove return value from View::handle
Similar to the rendering code, the event handling code is not expected
to fail. In fact, the main reason it could fail was because we called
into the rendering path, which could fail beforehand.
With the rendering being infallible we no longer require the event
handling functionality to account for error returns. Hence, this change
removes the 'Result' part of the return value from View::handle.