Use PathBuf instead of String for all things path related
We use file system paths in a couple of locations throughout the
program. However, so far we have always represented them as string and
passed that around.
Most if not all standard functions, however, work on std::path::PathBuf
objects. The back and forth conversion between the two types is
potentially costly and introduces an additional source for errors.
Hence, this change adjusts the code to only work with PathBuf objects.