Remove View abstraction
The abstraction of a view that we introduced, although conceptually
useful, hinders us too much in our anticipated design. Basically we have
to fit everything into this architecture but do not really need such an
architecture in the first place: we only have a single view and do not
anticipate more in the foreseeable future.
Hence, this change removes this abstraction and instead lets us use the
TermUi directly.