Avoid unnecessary redraw operations
When the selecting the top most or bottom most task and asking to change
the selection further up or down, respectively, we see a large amount of
flickering.
This flickering can be attributed to a redraw operation happening. Such
a redraw, however, does not need to happen as nothing has changed with
respect to what is being displayed (i.e., the selection is unaffected).
To this end, this change adjusts the UI select() method to indicate
whether the selection has actually been changed. By propagating up this
value we achieve the desired outcome of only redrawing when necessary.