Remove Orchestrator
The concept of a controller was introduced as part of the devised
model-view-controller architecture. This architecture got already
adjusted with the previous removal of the view. However, even the
concept of a controller that is, at least conceptually, usable by many
entities, does not pan out in Rust: We cannot simply share a controller
among multiple entities, because we need mutable access from each the
moment we want to save the current state.
To that end, this change removes the Controller trait and renames the
Orchestrator to Controller. From now on the controller is merely meant
to be an abstraction over the tasks, providing higher level
functionality on top. Since it is no longer intended to be a shared
entity, the UI now takes ownership of the Controller.