Introduce the concept of a renderer
This change introduces the concept of a renderer to the UI part of the
program. The Renderer trait is a manifestation of this concept. Objects
implementing this interface can be used for rendering the UI.
The introduction of this entity helps achieve a clean separation between
core logic and drawing related functionality. We use the gui crate's
Renderer trait because we are in the phase of transitioning all UI
related functionality towards using this crate and the remaining parts
of the UI will follow suit shortly.