Make TermUi struct the root widget
Until now we have used only chosen parts of the gui crate, specifically
the event functionality and the Renderer trait. With this change we
finally use the full infrastructure it provides. In particular, we use
an instance of the gui::Ui struct to manage all our widgets. With this
approach we are required to provide a root widget. The existing TermUi
is adjusted slightly to take on this role. Right now, it is the only
widget we have. Conceptually, we would also make the input/output area a
widget, but that is left for a follow up change.