Move update check into renderer functionality
As originally envisioned each widget would take care of keeping track of
whether it got updated since the last rendering procedure and only cause
a redraw if something changed. With the introduction of a dedicated
renderer entity this design was already found to be questionable. Also,
with the move towards having a dedicated widget for the input/output
area we can now have a completely standard implementation of the
Renderable::render method, as we no longer have to manually invoke the
corresponding render() method from the TermUi object).
Given these changes, this patch finally moves the update-condition check
into the renderer object itself. This has the nice effect of allowing us
to use the custom derive functionality for providing the standard TermUi
traits each widget has to implement, removing quite a bit of code
duplication.