Directly send events during testing instead of encoding them as a string
At time of its inception, the TermUi object required an object
implementing the Read trait and would provide the means to extract
termion events from it, which is why we have always used a string
representing the input events in the higher level functional tests. We
subsequently moved away from this design to the point that event
dispatching is happening completely outside of the TermUi object,
eliminating the need for this encoding.
Instead, this change adjusts said tests to directly dispatch event
objects as defined by the gui crate to the UI.