Add support for removing input via backspace key
We are pretty constrained in terms of how a user can input a particular
TODO item. In particular, we only support appending of characters, not
removal of any kind.
To fix this problem, this change adds support for the backspace key to
the program. That means we can now support input removal. We still have
no cursor or anything of that sort (and are not sure we need want/need
that) but this patch is a first step into the right direction.