Rework conditional update logic
In order to avoid unnecessary redraw operations and, with that,
potential flicker in the terminal, we employ a conditional update scheme
where we only redraw everything if something in the UI changed. To
accomplish this feat we have so far kept state that was internal to each
widget indicating whether it needs to be redrawn. This was known to not
work: terminals only support line based content clearing and so it is
simply not possible to just update a single widget conditionally. If we
cleared its area we would have cleared parts of another widget that
should have been unaffected as well.
This is a hard problem and we cannot easily solve it on the renderer
side of things. If, say, the root widget does not need to be updated but
a child does, the moment we get to the rendering of the child the root
widget's rendering code already executed (or not; but certainly can no
longer be invoked). That is because the root widget lies below other
ones and, hence, is rendered before them. So If we encounter a widget
that requires a redraw at this stage we are out of luck.
The solution to this problem is to effectively propagate the need for
update all the way up to the point where the rendering is triggered. To
accomplish this, we make use of the newly introduced event chaining
mechanism that allows us to conditionally add an update event to another
event.
With this infrastructure in place we now only invoke the rendering code
if the event handler indicated that an update operation is necessary.