Introduce the concept of a task ID
So far we have assumed an array of tasks that we can index into freely.
That was a reasonable assumption to make given the constraints we had.
In the future we plan on splitting the TermUi widget into multiple
widgets. Each widget may have its own view on a set of tasks.
Unfortunately, that means that we will somehow need to share the task
database among the various views. Because we want to occasionally
remove tasks, which requires mutable access, we cannot have multiple
references to the tasks object. And this is where it all breaks and we
cannot work with indices anymore: if a view decides to remove a task and
wants to do so by index, it needs to be sure that the index represents
said task. That cannot be guaranteed.
To that end, this change introduces task IDs. Such an ID uniquely
identifies a task. Instead of indices, all client facing task operations
will operate on such an ID in the future.