Implement task removal through an event
Once we split up the TermUi widget we will have the problem that we can
no longer directly access the controller object from the newly
introduced widgets -- we cannot share it in a mutable manner between all
of them. To that end, the controller will stay in the TermUi but the
TermUi will be able to be told what to do with it.
This change marks a first step into this direction in that it replaces
the direct task removal logic we had used so far with an event based
one. That is, a task can now be removed by sending an event specifying
the ID of the task to remove to the TermUi widget.