Make InOut widget focus the previous widget
This change adjusts the InOut widget's focus logic slightly: until now
we simply focused the parent after "closing" the widget. That works
while we have a single widget representing the tasks. We are currently
moving towards a model where multiple views on the tasks are possible,
each represented by its own view. In such a scenario we need actual
logic to set the focus back to the previously focused widget.
This change imports changes from the gui crate that allows us to do
exactly that and adjusts the InOut widget accordingly.