Introduce concept of a Query
With the plan to introduce multiple widgets referencing the same set of
tasks we run into the problem of how to share them effectively. We
cannot have true references as we need to be able to remove tasks or add
new ones while those widgets exist.
This change attempts to solve this problem by introduction of a new
type, Query, that represents an iterator over the task objects. Contrary
to ordinary iterators, a Query object has two advantages making it more
suitable for our use case:
1) Instead of holding on to a strong reference that would prevent
   additional mutable borrows from existing, we dynamically borrow the
   underlying task object.
2) To alleviate the risk for dynamic borrowing violations, the object
   provides an interface based on internal iteration as opposed to
   Rust's default external one. This way we do not have to leak borrows.