Introduce TaskListBox widget
With this change we introduce a new widget, TaskListBox, that takes care
of displaying the tasks we want to display in the form of a list box. By
factoring out this functionality we get a nicer separation of concerns
as well as the capability to create this widget multiple times.