Refactor custom event handling
The sending, receiving, and handling of custom events was not very well
designed so far. In particular, there were two major issues:
1) The flow of events was pretty convoluted. In particular, it was not
   easily possible to find the supposed sender of a received event just
   by looking at the code.
2) Events were defined in a very ad hoc fashion, making it not
   immediately obvious what was happening. E.g., a custom event
   containing a task::Id object is supposed to remove the task with the
   given ID.
3) Because of what effectively amounted to a repeated usage of Any
   within custom events, we lost out a lot on type safety.
4) The usage of Box::downcast_ref() which we used in order to nicely
   structure the code (not possible with Box::downcast()) prevented us
   from moving out of a box which in turn made the code employ
   unnecessary clone() operations.

With this change we introduce a new enum, TermUiEvent, that comprises
all events that are sent within the UI. That means the only allowed
custom event contains this enum in boxed up form. Further, we can now
get back to match'ing on such an event as we would usually do with
enums. Lastly, we get back pretty much full type safety.