Add support for changing the state of a task
Up until now a task was essentially just a summary which in turns is
just a free-form string. In order to be useful we will need to store
additional meta data along with this string.
This change marks a first step into this direction by adding a new field
representing its state. A task currently can be in two states: not
started and completed. The UI is adjusted to visualize those states
differently and can also be used to change the state.