Add support for 'g' and 'G' keys
We still have very limited task management and navigation shortcuts
available, making the program not as straight forward to use as it could
be. This change adds support for jumping to the beginning and end of the
task list. As one of the program's proclaimed design guidelines is to
to stick closely to well-known Vim key bindings, the chosen keys are 'g'
and 'G', respectively.