Introduce TabBar container widget
The next step in the evolution of the program is to support multiple
views on the set of tasks. The way this is going to happen is through
the creation of multiple TaskListBox widgets.
In order to arrange these widget this change introduces a container
widget that arranges those widgets in the form of tabs: a TabBar. An
object of this type is rendered as a single line at the top of the
terminal representing the individual tab names.