Remove parent_id and children from all widgets
Over a series of changes to the gui crate the 'parent_id' and 'children'
attributes of a widget got moved under the umbrella of the crates Ui
struct. Furthermore, the WidgetRef trait that was used previously got
removed as part of this refactoring.
This patch adjusts the program to adhere to those changes by removing
those members from its widgets.