Make Id infrastructure reusable
The 'unique_id' function we have for creating a Task's Id is not
accessible by other consumers but also flexible enough to cater to them
conceptually. In the future we anticipate the need to have similar Id
objects (although represented as different types) for bookkeeping in
other contexts.
This change factors out the Id related functionality into its own
module. Due to the need to create multiple Id namespaces, we add a type
parameter to the Id that separates different instances from each other.

Design-wise there a couple of things to note. In particular, the concept
of an ID allocator was briefly introduced, mostly in order to handle
persisted IDs, but overall not found to be workable.
In addition to allocating IDs, releasing and reserving them (if they
were loaded from a file) turned out to be not feasible. The mean reason
is that in order to implement releasing we probably want to track IDs on
an individual basis. What was explored was a range based allocator that
just keeps track of a range containing all the allocated IDs (that is,
with everything not inside of this range being "free"; this logic can
neatly be crammed into a single atomic variable). Naturally, with this
design is not possible to track individual IDs. Similarly, reserving
basically requires release functionality in order to flag duplicates.
Overall, this approach was not found to be workable and the concept of
persisted IDs (at least as initially devised) was abandoned.