Replace Deref impl for NamedTempFile with path() method
The NamedTempFile struct that we have created for testing purposes
implements the Deref trait to retrieve the path of the file. This trait
should generally only be implemented for smart pointers or similar
objects. In this case, it provides no benefit over an explicit method
call but just obfuscates things.
To that end, this change removes the implementation of this trait and
replaces it with the more explicit path() method.