Remove Tasks::FromIterator
We have got two ways to create a Tasks object from conversion: we can
create it from a vector or from an iterable. Both conversions were
initially devised to be used for testing purposes only, but have since
creeped into production code.
In the future the Tasks object will be more complex than a list of tasks
and, hence, no longer be trivially constructible from a set of tasks (in
one form or another). To that end, this change removes the FromIterator
trait implementation for the Tasks object and adjusts all client code to
use the vector version instead. The TermUi testing code in particular is
simplified by this step because it logically just has to work on a
vector of tasks and does not need a fancy interface on top.