Implement From<Vec<Task>> for Tasks only for tests
In the future the Tasks object will constitute more than a mere
collection of tasks. Hence, creation from a vector of Task objects will
no longer be possible.
We have already remove the FromIterator implementation for this struct
and with this change we define the From<Vec<Task>> conversion only for
testing purposes. Along with this change we also remove the
implementation of the Default trait, for the same reason stated above.