Provide alternative constructor for Controller object
The Controller struct currently can be constructed only by supplying
a path specifying a file with the data from which a Tasks object can be
constructed. That is a bit clumsy for testing code where we have to
create such a Tasks object just to write it into a file just to have the
Controller read it back. We might as well just pass in the Tasks object
directly.
This change adds a second constructor that opens up this possibility.