Implement Id based comparison for Task objects
Very early on, upon introduction of the Id field to the Task struct, we
implemented basic equality comparison by simply comparing the summaries
of the two tasks. That was not great, and it was only meant for testing
purposes, but by now we rely on equality comparison in a bunch of
places. Hence, we should fix comparisons to work properly.
To that end, this change corrects the PartialEq implementation for the
Task struct. To keep the behavior for the tests that rely on non-Id
based comparisons (for good reason) we introduce a new struct TaskVec
that preserves the previous comparison behavior.