Factor out task serialization & deserialization into separate module
In the future we will need to serialize and deserialize more complex
object structures (with interconnections among each other) than a simple
vector of tasks. We have already felt the complexity of having the
serialization and deserialization logic together with the "business"
logic of the program.
To cope with the anticipated increased complexity emerging in the
foreseeable future, this change separates those concerns cleanly by
moving the serialization and deserialization logic for tasks into a new
module, ser. The name was chosen as it is concise while also avoiding
collisions with serde, the serialization crate we use.