Provide serializable and deserializable ID
So far we have used IDs for identifying in-memory objects. That will not
be enough in the future for which we anticipate the need to persist
objects with IDs while requiring them to be deserializable again with
the same ID. Getting this right is not straight forward. The idea is to
have different ID types for the different purposes and to avoid mixing
in-memory and persistent IDs.
Pushing the needle some more towards this anticipated use-case, this
patch introduces a new ID type, ser::Id<T>, that can be serialized and
deserialized properly.