Add support for tags
Very early on one of the program's goals was to move into a direction
where arbitrary tags can be associated with tasks and to allow for more
sophisticated filtering of tasks by means of this infrastructure. This
change marks a first step into this direction. In particular, it
introduces the required infrastructure to serialize and deserialize
tags.
Persisting tags is not as straightforward as it may seem at first
glance. Specifically, conceptually a tag has two forms, the abstract
concept of it (e.g., wither a task is completed or not could be
represented as a tag 'complete') and a particular instance of it. For
boolean tags the difference is minor, but once we open them up to
contain one of many predefined values, such a view is essential. With
this change, these two concepts are represented as two structs, Template
and Tag. What complicates matters is that both have to be linked,
in memory but also in serialized form. We use the previously introduced
ID objects to accomplish this job. However, care must be taken to not
place too much trust into an ID that got read back into the program. In
particular, it is always possible that an ID we just read in actually
clashes with an ID we have already created.
In order to solve this problem in a holistic manner, we differentiate
between IDs as used during serialization and deserialization and those
used during runtime of the program. Now, the serialization ID is used
only to connect the tags as they were read back in. Once this connection
has been established, all IDs get replaced with freshly allocated
(collision-free) in-memory IDs.
This change only introduces the basic infrastructure for dealing with
tags. It does not hook them up to the UI and does not necessarily
provide a complete interface to all of its features.