Make sure a 'complete' tag is always present
With an upcoming change we would like to convert the existing task state
(which currently just indicates whether a task has been completed or
not) into a tag. This tag is sort of special, because it will be used by
the UI directly, similarly to how the state field is used already
nowadays. As such, it would be best if it were available unconditionally
and would not need to be defined by the user.
To that end, this change adds logic that adds this tag to the in-memory
Templates object. If it was present already because it got persisted, no
action will be performed. It also adds a reference to the template
representing this tag directly into the Templates object. This way we
can make sure not to have to search for this tag over and over again and
have to unwrap the result but instead have guaranteed access.