Create new tasks by supplying summary strings
In the past we were able to be very liberal when it comes to who could
create a task and what was required to do so. In particular, a string
was enough. That will not be the case any more in the future, where a
Task object needs to have an internal reference to the tag templates
available, such that we can keep our task centric view in the UI and
don't have to deal with other objects.
To that end, this change reworks task creation slightly: the existing
constructor creating a task form a summary is defined as being for test
usage only. The newly introduce 'with_summary' constructor (which will
gain more arguments in the future) is private.  Higher-level APIs are
adjusted to just hand the summary string down, instead of creating the
Task object themselves.