Clip all writes to the terminal according to bounding box
In its current form the program has the problem that overly long task
summaries wrap around in the terminal and cause confusing output. With
this change we make sure to properly clip overly long strings to be
displayed before writing them out to the terminal. This means that
overly long strings will be truncated silently so as to no longer cause
distortion in the output.