Set cursor position to selected task
When working in the terminal with a program such as tmux which allows
free copying of text using the keyboard, complications arise when the
cursor is located at a more or less arbitrary location when selection
mode is entered. That is basically what is the case as it stands: the
cursor ends up being located wherever the last 'write' operation ended
on.
This change makes things a bit more deterministic. In particular, we
remember the location of the selected task and set the cursor to the
beginning of its first character. That allows for very nice
selectability.
Note, however, that strictly speaking the order in which widgets are
rendered may change and that may break this functionality.