Adhere to change of return value of Ui::handle
The Ui's handle method recently changed its return value from an
optional UiEvent to an optional MetaEvent. This change was performed to
make event handling code easier controllable by returning all unhandled
events instead of just the last one. For us that means that we now need
to iterate over a potential event chain to deduce whether to redraw
everything or to exit the program.
Along with that the crate also moved from passing an Any to a Renderer's
render method to now passing an actual Widget trait object.
This patch adjusts the program to adhere to these changes.