Correctly clear input/output area state after every key press
This change fixes a bug caused by the introduction of the InOutArea
widget. Before having a dedicated widget representing the input/output
area we only had one widget overall. With that architecture we were
easily able to clear the in/out area on every key press, simply because
we received all key events.
With the introduction of a dedicated widget that was no longer easily
possible, because there was no single entity through which all events
passed. Hence, the state of said area would not stay the same until the
widget was instructed to update this state.
This patch makes use of the event hooking functionality recently
introduced to the gui crate, which allows a widget to snoop all incoming
events. Doing so, it the InOutArea widget can reset its own state as
desired, provided it is not currently focused (in which case text is
actively input).