Introduce custom event unwrapping helper trait
With the three levels of events that the gui crate provides it can be
quite cumbersome to retrieve an expected value from an event. Our
testing infrastructure in particular makes use of custom events in many
locations and unwrapping them every time to check for correctness causes
a lot of boilerplate code.
To alleviate the pain this patch introduces a new trait, CustomEvent,
that is used by the testing infrastructure to more comfortably unwrap
custom events. In the future we anticipate to use this trait more often.