Add tests for InOutArea widget state
In order to avoid regressions in the realm of the InOutArea widget, this
patch adds tests that verify the state the widget should be in after
certain interactions.
The change also establishes the usage of new TermUiEvent branches
dedicated to returning events in response to a request. The TermUi
widget takes precautions to ensure that TermUiEvent objects are the only
custom events that pass through it. By returning, say, a vector of Task
objects directly we violate that constraint. That worked in the past
because we were returning this vector from the TermUi widget itself, but
since we are now redirecting a request to the InOutArea widget, which in
turn will create an event that passes through the TermUi, we truly are
not playing by the rules.