Move TaskListBox management into the TabBar widget
The TabBar widget was devised as a container widget for the various
widgets making up the individual tabs. However, since we initially did
not have a way to make use of various tabs (we have no way to split up
all tasks into subsets that can be displayed individually), we just
created a single tab filled by the TaskListBox that displays all tasks.
The orchestration of the creation of all those widgets was still
orchestrated by the TermUi object.
With this patch we change matters slightly, in that we move the creation
of the TaskListBox widget into the responsibility of the TabBar.