Make InOutArea remember the previously focused widget
Whenever the InOutArea widget was done inputting text it would logically
transfer control back to the widget that had asked it to input text in
the first place. In our case, transfer of control pretty much boils down
to changing the input focus. In order to refocus the widget from which
control was transferred we relied on the gui crate to remember the
widget that previously had the input focus.
With this change we change this approach, and in order to understand why
that is the case a little bit of context is in order: Due to isolation
of widgets from one another a widget may not necessarily know the ID of
another widget that is supposed to receive an event. Because of that a
widget cannot necessarily address said widget directly. So instead, we
just emit this event, let it propagate to the root widget (assuming the
recipient is not a parent, in which case the event would be handled
directly), and the root widget will have enough knowledge to know who
can handle this kind of event. So it will forward the event to the
widget it judges most suitable. From there each widget has a little more
knowledge and can reroute the event until it finally reaches its
destination.
With this explanation in mind, we adjust the way the TaskListBox widget
indicates a change in state to the InOutArea according to the outlined
solution: the list box will merely emit an event, it will reach the root
widget which will redirect it to the InOutArea widget which it has
knowledge of because it is one of its children. Because of this
restructuring the TaskListBox no longer needs direct knowledge of the
InOutArea widget, which decouples the components a bit more.