Always select a newly added task
One major annoyance when working with the program was that the selected
task would not change when adding a new task. That is particularly
confusing when the list of tasks cannot be displayed on the screen,
i.e., if scrolling is involved, in which case seemingly nothing changed
for the user.
With all the preparatory changes made previously it is not possible to
fix this problem. Instead of having the task list box adjust its input
focus in a way that anticipates the appearance of a soon-to-be-added
task we now send back an event once the task has been added and let the
widget adjust its selection in such a way as to select said task.