Make Query objects filterable with tags
The main reason provided for the introduction of the Query struct was
that it allowed us to have a reference to a Tasks object from many
objects. However, that was not the full story. The concept of a query
also supports the notion of filtering by tags. That is, a query can have
certain tags attached to it that must (or must not) be present in order
for a task to appear during iteration.
This change implements exactly this functionality. With it, a Query
contains a "formula" describing the tags that are desired. Such a
formula is comprised of conjunctions (AND) and disjunctions (OR) of tags
and is stored in conjunctive normal form (CNF). This form simplifies our
internal processing significantly but it also influences the API:
conjunctions always associate with the most recent conjunction (and
would start a new set of disjunctions) and so some preprocessing may be
necessary to represent the intended meaning.
Query objects are now created through a QueryBuilder. This object got
introduced to keep queries themselves immutable. The API seemed
unnatural if it allowed for custom modifications at runtime. In the
future a Query is envisioned to represent a view onto the set of tasks
and it is not meant to be modified permanently after its creation -- a
new Query object should be created instead.