Enable the cursor when inputting text
So far we have disabled the cursor during program startup and enabled it
again on shutdown. The reason is that the cursor is a mere distraction
when rendering widgets in a UI. When inputting text, however, things
look different and having a cursor is what is intuitively expected. To
that end, this change enables the cursor when rendering the InOutArea
widget and if it is currently in input mode.