Refactor input event handling in InOutWidget
This change refactors the handling of input events in the InOutWidget
slightly. In the future we plan to add more cases in which case the
individual matching of the in_out state field in every branch introduces
way too much syntactic overhead. So move this part out of the individual
branches towards the top.