Add input index handling to InOutArea widget
A previous change has introduced very basic support for showing a cursor
when inputting text. To take this functionality one step further this
change adjusts the InOut area state object to also contain the input
cursor position when in Input mode. The renderer is adjusted accordingly
to use this index as the cursor position as opposed to having it be
placed after the text unconditionally as was done beforehand.