Do not allow empty task additions
So far we have simply allowed additions of tasks with an empty summary
from the UI. However, there really is not much of a point doing so.
Hence, with this change we disallow the addition of empty tasks through
the UI. It is still possible for a user to edit the JSON tasks file,
however, and there is no need to prevent that case as it is completely
outside of the intended usage of the program and not harmful.