Rename Controller struct to State
This change renames the Controller struct to State. The Controller was
originally introduced as part of a design that early on turned out to be
infeasible to implement. Yet, the struct itself never got in the way of
implementation matters and hence, stayed until now. In fact, we always
wanted the one level of indirection the object provided. The main
problem was that the exact set of responsibilities those objects would
have was mostly undefined.
By using the arguably more explicit name State we hopefully resolve most
ambiguities surrounding this struct: its main task is to manage and
contain the majority of the program's state. This state includes the
task state that already falls under its responsibility, but will be
enlarged with future changes.