Redesign Tasks & State responsibilities
One of the main goals of the newly introduced State struct is to manage
the state surrounding tasks. Part of this state are the tag templates,
which so far were owned by the Tasks object. With this change we
transfer said ownership into the State struct. In the future we will
need those templates outside of a Tasks object as well, and this
refactoring helps enable this use case.
Furthermore, Tasks objects arguably contained too much logic surrounding
the loading of tasks. So with this change we move the file management
over into the state functionality. The validation of tags when loading
from a serialized state, on the other hand, is moved into the Tasks
struct, in slightly changed form.