Give Query objects a name
In the not-so-distant future we want to visualize queries and they tasks
they represent in individual tabs of the TabBar widget. In order for
those tabs to have any significance to the user they should contain some
sort of identifier.
With this change we make Query objects contain a name attribute that
will be used as said identifier. Strictly speaking a query is just an
abstract entity and there is no need to name it, so the titles of the
individual tabs should be stored elsewhere. However, in the short term
future having a Query contain a name is the most reasonable way forward.